test: move func tests to tests/func
