+++ tests/func/test_analytics.py
