+++ tests/func/test_cache.py
