+++ tests/func/test_cli.py
