+++ tests/func/test_config.py
