+++ tests/func/test_diff.py
