+++ tests/func/test_gc.py
