+++ tests/func/test_ignore.py
