+++ tests/func/test_install.py
