+++ tests/func/test_lock.py
