+++ tests/func/test_metrics.py
