+++ tests/func/test_pkg.py
