+++ tests/func/test_remove.py
