+++ tests/func/test_stage.py
