+++ tests/func/test_tag.py
