+++ tests/func/test_updater.py
