+++ tests/func/test_utils.py
